+++ formula-android-tests/src/test/java/com/instacart/formula/FragmentLifecycleTest.kt
@@ -8,6 +8,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
 import com.instacart.formula.android.Feature
 import com.instacart.formula.android.FeatureFactory
+import com.instacart.formula.android.FragmentFlowStore
 import com.instacart.formula.android.ViewFactory
 import com.instacart.formula.test.TestFragmentActivity
 import com.instacart.formula.test.TestFragmentLifecycleCallback
@@ -36,7 +37,7 @@ class FragmentLifecycleTest {
                         contract = TestLifecycleKey()
                         initialContract = contract
                     },
-                    contracts =  {
+                    fragmentStore = FragmentFlowStore.init {
                         val featureFactory = object : FeatureFactory<Unit, TestLifecycleKey> {
                             override fun initialize(dependencies: Unit, key: TestLifecycleKey): Feature {
                                 return Feature(