+++ formula-android/src/main/java/com/instacart/formula/android/internal/FeatureBinding.kt
@@ -6,7 +6,7 @@ import com.instacart.formula.android.FragmentKey
 /**
  * Defines how a specific key should be bound to its [FeatureFactory]
  */
-class FeatureBinding<in Component, Key : FragmentKey>(
+class FeatureBinding<in Dependencies, Key : FragmentKey>(
     val type: Class<Key>,
-    val feature: FeatureFactory<Component, Key>,
+    val feature: FeatureFactory<Dependencies, Key>,
 )