+++ samples/stopwatch-compose/src/main/java/com/instacart/formula/compose/stopwatch/StopwatchApp.kt
@@ -4,6 +4,7 @@ import android.app.Application
 import android.util.Log
 import com.instacart.formula.FormulaAndroid
 import com.instacart.formula.android.FragmentEnvironment
+import com.instacart.formula.android.FragmentFlowStore
 
 class StopwatchApp : Application() {
 
@@ -20,7 +21,7 @@ class StopwatchApp : Application() {
             activities = {
                 activity<StopwatchActivity> {
                     store(
-                        contracts = contracts(Unit) {
+                        fragmentStore = FragmentFlowStore.init(Unit) {
                             bind(StopwatchFeatureFactory())
                         }
                     )