+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 - Enable fine-grained control of dispatching via `Plugin.defaultDispatcher`, `RuntimeConfig.defaultDispatcher`, `Transition.ExecutionType` and `Effect.Type`.
 - Remove `RxStream` type alias.
 - Replace `implementation` function with a value
-- Removed `FlowFactory` and `Flow`
+- **Breaking**: Removed `FlowFactory` and `Flow`
+- **Breaking**: Replace `FragmentStoreBuilder` with `FeaturesBuilder`
 
 ## [0.7.1] - June 28, 2022
 - **Breaking**: Rename `FragmentBindingBuilder` to `FragmentStoreBuilder`